+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Braintree iOS SDK Release Notes
 
-# unreleased
+## unreleased
 * Attach zip file of static Braintree xcframework modules to each GitHub release
+* BraintreeCore
+  * Fix bug where instantiating multiple feature clients with  universal links in the same view causes `handleOpen` to return in the wrong feature client
 
 ## 6.29.0 (2025-02-24)
 * BraintreePayPal