+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Attach zip file of static Braintree xcframework modules to each GitHub release
 * BraintreeCore
   * Fix bug where instantiating multiple feature clients with  universal links in the same view causes `handleOpen` to return in the wrong feature client
+* BraintreePayPal
+  * Add `BTContactPreference` to set contact information section preference within the payment flow
+  * Add `contactPreference` property to `BTPayPalCheckoutRequest`
 
 ## 6.29.0 (2025-02-24)
 * BraintreePayPal